requirements.txt
ipython==8.15.0
openai==1.8.0
Pillow==10.2.0
PyYAML==6.0.1
Requests==2.31.0
streamlit==1.30.0
streamlit_authenticator==0.2.3
load_dotenv
azure-ai-vision-imageanalysis
azure-ai-documentintelligence